Fix map nav icon for roundabouts
navd helper uses bannerinstructions maneuver modifier for type of icon to show for upcoming instructions. MapBox shows a right turn modifier (LHD) for all roundabouts as the vehicle is technically exiting this direction from the roundabout.
For navd helper to show the correct type of turn for the roundabout, we can use the degrees of the roundabout turn that Mapbox supplies. This can be then used to show better instructions for the driver. 
For LHD vehicles in degrees, 
0-67 : sharp right
67-112 : right
112-157 : slight right
157-202 : straight
202-247 : slight left
247-292 : left
292-360 : sharp left