navd: allow the mapbox host to be configured via an environment variable
This seems to be the only hard-coded Comma host in the code. Athena and the API can be set with `ATHENA_HOST` and `API_HOST` variables, as seen at:

- https://github.com/FrogAi/FrogPilot/blob/c9289bef84aab5105da740ffd6c40f3f41530759/system/athena/athenad.py#L44

- https://github.com/FrogAi/FrogPilot/blob/c9289bef84aab5105da740ffd6c40f3f41530759/common/api/__init__.py#L8

My usecase is that I have a self-hosted Comma-API-compatible server (https://github.com/USA-RedDragon/connect-server) that  essentially acts like the official Comma API. This all works and is nearing a proper release, however navigation is hard-coded to `maps.comma.ai` on devices, so routing is a no-go without a change like this.