Increase Allowable QLog Size for Uploader
This is cherrypicked from commaai/openpilot@c287232

The uploader limits the size of uploaded logs.  The size is calculated prior to compression, prior to a Nov 25, 2024 change this was set at 5MB.

Comma has since increased this to 25MB since a potential bug could be one that causes very large log sizes.

If a log exceeds this limit, the uploader outputs an error, but that error is not logged anywhere and would only be seen if tmux was running at the time the error occurred.  The log is then marked uploaded and never tried again.

When this happens the file list in useradmin shows missing qlog files.

In my case, my logs were bloated because these logs contain a `git diff` output and all of the compiled translation files were appearing in the diff leading to about 1.75MB of additional logging details.  I will update my .gitignore file to fix this, but it seems reasonable to increase this limit if comma is as well and their logic for doing so seems rational.